moved message events to cog
