+++ src/beefcmd/cogs/events_listener/member_events/member_join.py
@@ -1,0 +1,19 @@
+import discord
+from guilds import read_guild_log_channel
+from datetime import datetime
+
+async def member_join_event(bot, member):
+    channel_id = await read_guild_log_channel(member.guild.id)
+    channel = await bot.fetch_channel(channel_id)
+    await channel.send(embed=await join_message_embed(member, bot.user.avatar.url, member.guild.name))
+    
+async def join_message_embed(user: discord.Member, icon_url, guild_name):
+        joinembed = discord.Embed(title=f"Hello!!!!!!!!", color=discord.Color.green())
+        joinembed.set_thumbnail(url=user.avatar.url)
+        joinembed.add_field(name="", value=f"<@{user.id}> **joined**", inline=False)
+        joinembed.add_field(name="", value=f"Account created: `{user.created_at.strftime('%d/%m/%Y')}`", inline=False)
+        joinembed.add_field(name="", value="", inline=False)
+        joinembed.add_field(name="", value="", inline=False)
+        joinembed.set_author(name="Beefstew", icon_url=icon_url)
+        joinembed.add_field(name="", value=f"{guild_name} - {datetime.now().strftime('%d/%m/%Y %H:%M')}")      
+        return joinembed
